[RUI-241]: Move getScatterChartProOptions to utils file
