+++ src/components/charts/scatter/ScatterChartDefaultPro/ScatterChartDefaultPro.utils.ts
@@ -1,14 +1,85 @@
-import { ChartData } from 'chart.js';
+import { ChartData, type ChartOptions } from 'chart.js';
 import { DataResponse, Dimension, Measure } from '@embeddable.com/core';
 import { Theme } from '../../../../theme/theme.types';
 import { getThemeFormatter } from '../../../../theme/formatter/formatter.utils';
 import { getChartColors } from '@embeddable.com/remarkable-ui';
-import type { ChartPointClicked, ScatterChartInputPoint } from '@embeddable.com/remarkable-ui';
+import type {
+  ChartPointClicked,
+  ScatterChartInputPoint,
+  ScatterDatasetWithOriginal,
+} from '@embeddable.com/remarkable-ui';
 import { getDimensionMeasureColor } from '../../../../theme/styles/styles.utils';
 import { getDimensionFieldName } from '../../../../utils/data.utils';
 import type { PointClickArgs } from '../../charts.types';
 
 export { getDimensionFieldName };
+
+export const getScatterChartProOptions = (
+  { xMeasure, yMeasure }: { xMeasure: Measure; yMeasure: Measure },
+  theme: Theme,
+  noValueLabel: string,
+): Partial<ChartOptions<'scatter'>> => {
+  const themeFormatter = getThemeFormatter(theme);
+
+  const formatValue = (measure: Measure, value: number | null | undefined): string => {
+    if (value === null || value === undefined) return noValueLabel;
+    return themeFormatter.data(measure, value);
+  };
+
+  return {
+    scales: {
+      x: {
+        ticks: {
+          callback: (tickValue) => {
+            const v = typeof tickValue === 'number' ? tickValue : Number(tickValue);
+            return themeFormatter.data(xMeasure, v);
+          },
+        },
+      },
+      y: {
+        ticks: {
+          callback: (tickValue) => {
+            const v = typeof tickValue === 'number' ? tickValue : Number(tickValue);
+            return themeFormatter.data(yMeasure, v);
+          },
+        },
+      },
+    },
+    plugins: {
+      tooltip: {
+        callbacks: {
+          label: (ctx) => {
+            const ds = ctx.dataset as ScatterDatasetWithOriginal;
+            const orig =
+              ds.originalData?.[ctx.dataIndex] ??
+              (ctx.dataset.data[ctx.dataIndex] as
+                | { x: number | null; y: number | null }
+                | undefined);
+            const prefix = ds.label ? `${ds.label}: ` : '';
+            if (!orig) return prefix;
+            return `${prefix}(${formatValue(xMeasure, orig.x)}, ${formatValue(yMeasure, orig.y)})`;
+          },
+        },
+      },
+      datalabels: {
+        labels: {
+          value: {
+            formatter: (_value, context) => {
+              const ds = context.dataset as ScatterDatasetWithOriginal;
+              const raw =
+                ds.originalData?.[context.dataIndex] ??
+                (context.dataset.data[context.dataIndex] as
+                  | { x: number | null; y: number | null }
+                  | undefined);
+              if (!raw) return '';
+              return `(${formatValue(xMeasure, raw.x)}, ${formatValue(yMeasure, raw.y)})`;
+            },
+          },
+        },
+      },
+    },
+  };
+};
 
 export type CellValue = string | number | boolean | null | undefined;
